gemini_semgrep_miso_copy/requirements.txt
requests 
google-genai 
PyGithub 
gitpython 
